requirements.txt
python-telegram-bot==20.7
youtube-search-python==1.6.6
requests==2.31.0
beautifulsoup4==4.12.2
urllib3==2.0.7
python-dotenv==1.0.0
aiohttp==3.9.1